Remove special treatment of constant assignment (#107)
Simplify the special treatment of statements like `x = 10`.  This statement will be treated as shorthand for `x = op.Constant(10)`.

The more complex treatment in-place currently is to allow the use of such a constant `x` as either an ONNX (tensor) value or an ONNX attribute value later on. The more complex handling may be a useful feature, but it is incomplete currently. Simplifying the code for now. The feature can be added later if desired.