Add typenames and opsets to top level namespace (#198)
Add typenames and opsets to top level namespace. These will be commonly
used by users, so it is better to have it in the top level namespace. It
also avoids users taking a dependence on internal file/sub-module
organization.

Once this is accepted, I can also update the documentation/examples to
reflect this. (Updated just the README for now.)

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>