Update on "feat(atenlib): atenlib function registry"
Registry for all atenlib functions, which supports overloading. This change implements the `torch_op` decorator for registering all functions in the ATen lib so they are discoverable. It is not designed to be used by users, but rather as an info container so code gen knows which ops are implemented.

[ghstack-poisoned]