feat(atenlib): atenlib function registry (#260)
Stack from [ghstack](https://github.com/ezyang/ghstack) (oldest at
bottom):
* __->__ #260

Registry for all atenlib functions, which supports overloading. This
change implements the `torch_op` decorator for registering all functions
in the ATen lib so they are discoverable. It is not designed to be used
by users, but rather as an info container so code gen knows which ops
are implemented.

`torch_op` also compiles the functions into OnnxFunction.