feat(atenlib): index_select; trace_only ops (#274)
Implement `aten_index_select`; enable tests for transpose.

This change also added the **`trace_only`** argument to mark functions
as trace only to work around functions that cannot be compiled. The
logic can then be tested, but the argument should be removed later when
onnxscript extends support for the syntax.