fix(atenlib): test split and support Sequence input (#528)
This PR adjusts the test code to be able to handle the output of
`torch.split` and adds support for sequence as input. Note that this is
different from the exporter's concat for symints, because at the point
when we have inputs processed and still get list of Tensors as a single
argument, we know it should be an ONNX Sequence tensor in the graph.

We also need to keep the addition in here and not in the exporter to
match behavior with onnxscript eager: we accept a list of tensors as
input when the op takes a Sequence.