Update base for Update on "Auto OpSchema for trace_only functions | feat(op_schema)"
This PR implements auto OpSchema generation for trace_only function as well. It leverages onnxscript's converter to create the same function IR as we do in `OnnxFunction`, but without translating the body.

We created an `TraceOnlyFunction` class to expose a similar interface as `OnnxFunction`. We may be able to define a Protocol `HasOpSchema` for both classes.

The change is tested on all the existing trace_only functions.

Fixes #630

[ghstack-poisoned]