Update base for Update on "Retire `get_schema` in Op | chore!(api)"
Remove `get_schema` in Op because (1) if an op does not have schema, it will not be in its opset either. (2) If an op does have schema, we don't need to consult with the opset. This change ensures a single correct way of accessing OpSchema.

Rename `opschema` to `op_schema` for naming consistency. 

[ghstack-poisoned]