Indexing and slicing support (#711)
Part 1 of extending the support for indexing and slicing (expressions of
the form `e[i_1, i_2, i_3]`).

This updates only the runtime (eager) component. The updates to the
converter will be part 2 (once this is accepted/finalized).

Primarily extends support to the use of non-scalar tensors as an index
(which is effectively a `Gather` operation). Also implements a specific
heuristic to choose between the use of `Slice` and `Gather` in some
cases where both can be used.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>