Fix op (MaxPool) | feat(torchlib) (#757)
Found this bug when I try to apply this op in a model
(https://github.com/pytorch/pytorch/pull/102773). Basically, if the
input padding is 3d, we still need to double it to fit the ONNX
requirement that we need begin and end padding. I am not sure why we
didn't catch this in ops_test.py though, it seems should be caught...

---------

Co-authored-by: Jay Zhang <36183870+fatcat-z@users.noreply.github.com>