Indexing and slicing support (part 2) (#827)
Fixes to the translation of expressions of the form `exp1[exp2]`. This
is a continuation of https://github.com/microsoft/onnxscript/pull/711
which fixed the eager-mode handling of such expressions. This PR
implements the same for the script translator.

In the expression `e[i_1, i_2, ..., i_n]`, `n` is either the rank of the
input tensor or any value less than that. Each index-value `i_j` may be
a scalar value (a tensor of rank zero) or a higher-dimensional tensor or
a slice-expression of the form `start:end:step`. Semantically, a
slice-expression `start:end:step` is equivalent to a 1-dimensional
tensor containing the corresponding sequence of values.

Main changes:
* Previously the translation of `X[Y]` assumed `Y` was a scalar. This PR
handles the situation correctly when the rank of Y is unknown.
* Handle default start/stop values using minint/maxint.

TODO:
* Doesn't yet handle default-values for start/stop when the
step-direction is unknown. This can be handled by emitting code to
choose the default value based on a runtime check on the sign of the
step-value.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>