Implement `aten::div.Tensor_mode` | feat(torchlib) (#988)
`aten::div.Tensor_mode` is implemented with two ONNX functions. When
`rounding_mode` is `None`, we use `aten_div`. When it is not None, we
use `aten_div_mode`. This way we don't need to handle when
`rounding_mode==None` in `aten_div`.

For `float16` inputs we need to cast to float32 to preserve precision.
Otherwise `-inf` sometimes becomes `inf` in the output.

- Additionally registers aliases "aten::divide", "aten::true_divide" to
`aten_div`.
- Supports saving mismatches to error reports
- xfail and documents off-by-one errors with float16
(https://github.com/microsoft/onnxscript/issues/990,
https://github.com/microsoft/onnxscript/issues/989)

Fixes #980