Implement onnx reference evaluator as an eager backend | evaluator (#383)
Stack from [ghstack](https://github.com/ezyang/ghstack) (oldest at
bottom):
* __->__ #383

In this PR

- Implement onnx reference evaluator as an eager backend
- Refactor some code in evaluator

I tested locally by replacing the ort evaluator with the reference
evaluator. Most tests passed and some tests failed. I did not want to
add xfails or bring the change to the torchlib tests yet. Would be good
to merge as-is so that we at least have an alternative backend
implemented.

Fixes #209