Move the test steps out of dev-release | ci(release) (#1030)
Move the test steps out of the dev-release pipeline because the tests
may fail due to dependency changes. For the dev pipeline we need to
ensure daily releases are successful and are less worried about failing
tests. The tests steps are moved to only exist in the official release
pipeline.