Handle name conflict (attribute parameter) when converting back to python (#1166)
Attribute parameters and normal values (like tensors) use the same
namespace in onnxscript/python. This can cause a conflict when
converting ONNX proto back to onnxscript (since these are different
namespaces in ONNX). Examples of where this happens shown in test-case
below: eg., when an attribute-parameter "yyy" is used as a value, the
onnxscript translator introduces a value called "yyy" which is bound to
the attribute-value "yyy".

Fix this in the exporter.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>