Simplify linked list design with a hash table (#1349)
Use a dictionary to map the values to the boxes in the doubly linked
list implementation to

1. Remove the private pointer from Node to LinkedBox
2. Simplify membership checks
3. Ensure elements in the list are unique

Also changed:
1. The node <-> graph relationships are now managed by Graph directly
and not by a property_modifier function. Meaning the user is responsible
to remove nodes from other graphs before adding the same node to a new
graph. This helps separate the concern of the graph from the linked
list, allowing us to treat the linked list as a normal list. The list
also doesn't need to handle exclusivity, meaning the same value will not
be removed from the original list when being added to a new list and can
appear in multiple lists, which is what we expect how lists behave.