[IR] Assign default names to nodes and values (#1371)
Create the `NameAuthority` class to allow graphs to assign default names
to nodes and values if they do not have one.

The class does not keep track of the names it has given or concerns with
any naming scopes. So it is possible to generate names that conflicts
with existing names. It is the responsibility of the user to ensure that
the names are unique (typically by running a name-fixing pass on the
graph).

We do not need to ensure name uniqueness in the IR because nodes and
values are all stored by reference. It is not possible to confuse one
value with another if they share the same name.