[CI] Fix test path to run tests/ in CI (#1377)
Run tests under the `tests/` directory. Previously we migrated `tests/`
from `onnxscript/tests` but I did not realize nox was not properly
configured to discover them. cc @BowenBao

1. Fix a serialization / deserialization issue with shape/type in the IR
2. Fix an import error in test
3. Stop running doctests by default and instead specify it in CI,
because when doctest is importing a module that has functions not
supported by onnxscript (in this case under `opt_output.py` which is
delayed imported in tests), import will fail.
4. Bump ORT to 1.17 to reduce memory leaks in CI.