Update example (#1521)
The example in `examples\pattern_rewriting.py` was not updated when the
pattern-matchers API were unified. Fix this.

Question: do we want this standalone example? I think the documentation
folder examples are a better place to add anything we want, and this can
be removed. (But leaving it here for now.)

TODO: Improve the logging info produced in verbose mode, especially for
the pattern graph.

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>