[torchlib] Add the identity nodes back (#1703)
In the modularization pass in the exporter, a single node like `clone`
can be lifted as a function. If we remove the only Identity node the
lifted function will have no nodes. This violates the ONNX standard.

Since removing identity nodes is fast, we are safe to include these
identity nodes in the torchlib.

onnxscript/tools/transformers_models/phi_test.py broke after #1613, it
is fixed by this change.

---------

Signed-off-by: Xavier Dupre <xadupre@microsoft.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>