Add test case for unsupported input-variable reuse in multi-output pattern matcher (#1731)
The multi-output pattern does not support a useful case, where the
pattern's input variables are reused elsewhere.

This PR adds a test-case illustrating the scenario. I was hoping that
some changes in the in-progress PR #1636 by Xavier might fix this. But I
don't think it does. Adding the test-case for now. Need to figure out
how to support this.

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>