Extend basic matcher to handle multiple-output-nodes (#1734)
This PR extends the basic matcher to handle multiple output nodes. This
provides an alternative to the generic-matcher algorithm, which is
incomplete and fails in some circumstances. This can also be useful in
debugging match-failures (when it is unclear if the failure is valid or
due to limitations of the matching algorithm). The drawback is that this
algorithm can, in some cases, be expensive, especially when the number
of output-nodes is large and the graph size is large. (So far, however,
we haven't encountered patterns with more than 2 output-nodes.)