Track nodes added in pattern builder (#1770)
This fixes a couple of issues with the graph pattern builder by
explicitly tracking the nodes added/created. This ensures that node
ordering is exactly the same as what the user specifies (which helps
with debugging and the verbose logs). In addition, we use a context
manager to track the nodes added via the use of overloaded operators
like + and *.

This also impacts how the "commuted" GraphPatterns are constructed. This
has also been cleaned up (allowing commute to handle multiple output
nodes as well).

Remove the unused "onnxop" imports left over after some refactoring a
while back.

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>