[IR] Remove ir.Attr* classes (#1817)
Remove ir.Attr* classes because they are used for isinstance checks,
which is not robust. This is because an ir.Attr can take any type and
should also be treated as that type of attribute when the type is set.
Using `isinstance` in this case will miss the attribute. To avoid
confusion and prevent incorrect usage, I removed the classes but kept
them as convenience functions for backward compatibility.