[IR] Fix external tensor path handling (#1823)
Previously the external tensors are deserialized with the base path and
the relative location combined. This is in accurate because we lose
information on which is the actual "location" that should be written to
the proto.

With this change, I added a new `location` parameter/attribute to
represent the ONNX proto relative "location", and update the `path`
attribute to be computed by joining the base path with the relative
location. As such users can always access the `tensor.path` attribute to
get to the data file. Updated `serde` to make use of this attribute.

@shubhambhokare1 please also update the offload pass in #1796 to make
use of the `path` attribute.

## BC breaking

The first parameter of the `ExternalTensor` initializer is renamed to
`location`.