[torchlib] Fix floor_divide
Implement floor_divide for int inputs. We implement it only for positive inputs (using integer division) because that is the usual intended case and is the most efficient.

Create op info for the aten op because the original op info does not produce the same input expected nor has the same behavior.

Mark traceable.