A few fixes relating to constant propagation (#1892)
Fixes a few different issues. Helps resolve an issue relating to
ir-based optimization for the Blender model in the benchmark.

* Move the utility for evaluating `Constant` op into the IR, and make
`const_value` automatically perform the related computation.
* Eliminate the dependence on the reference-implementation for
evaluation of Constant op.
* There are still a couple of issues relating to the use of
reference-implementation (eg., when we have tensor-valued attributes in
external-data format, and the use of float16) which will need to be
addressed separately, but the above bypasses this issue for Constant op
(and the Blender model).
* Make the optimizer robust to external-data-tensors whose files are not
available.