A couple of optimizer and rewriter extensions (#1937)
A few extensions motivated by ongoing transformer fusion optimizations:

Pattern matching:
* Extend pattern-matching pattern to allow specifying that extra-inputs
are allowed.

Optimizations:
* Concat (x) can be replaced by Identity(x)
* Redundant cast optimization was missing in core optimizer (though
present as a llama rewrite rule).
* Dropout optimizations moved into core optimizer (from rewrite rule;
rewrite rule has an issue, use of attribute instead of input, and it
seemed better to move it into core optimizer).

In general, for optimizations involving a single node, the core
optimizer is a better place (at least, as long as they are generic, and
not backend-specific) than rewrite rules. It is more efficient.

* Fix input/output size limit of constant-folding to be number of bytes.
(It is currently inconsistent, as number of bytes for one and number of
elements for another).