[IR] Implement register_initializer (#1941)
Implement `register_initializer(value)` on `Graph` for robustly adding
an initializer to the graph. Users can also directly modify the
`graph.initializers` dictionary, but this method does more comprehensive
checks before adding, and calling this method is simpler than modifying
the dictionary.

The method could be in the `convenience` too, but I put it here due to
its relevance and for better discoverability.