Handle input initializers correctly in constant folding (#1944)
Values that are both inputs and initializers of a model/graph should not
be treated as constants (and cannot be used for constant-folding).
Unfortunately, the single `const_value` field is class Value is used
both to indicate constant-values of proper constants as well as
initializer values of initializers. Ideally, the IR should provide an
easy way to distinguish this at the value level (with either an extra
boolean flag to indicate the value is an input-value or by using
distinct fields for "initializer_value" and "const_value".

Meanwhile, this PR introduces a workaround to handle the main issue.