Extend onnx to script converter (#1987)
Extend onnx to script converter to suppress initializers in generated
script, and replace them with randomly generated weights. This allows a
compact onnxscript representation of models with large initializers,
especially when we don't care about the exact weights. This is useful
for generating source-based test-cases from models.

Eg., this allows us to generate test-cases such as the one
[here](https://github.com/microsoft/onnxscript/blob/ca11a20625fb3b170c23798e03b4366d17d6bc1c/onnxscript/rewriter/onnxruntime/xformers/_smollm_1layer.py#L18).

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>