A couple of extensions to rewriter (#2001)
Extends the rewriter with a couple of features:

* A debugging mode to perform the pattern matching (without any graph
modifications) and to report instances that get the best score for a
match (even if incomplete). Helps quickly identify causes for mismatches
when we expect a match.

* Rewrite-rules can now specify a pre/post visitor method called before
applying it to a graph/function. This is useful for rules that need to
create "cached" values that are reused across multiple instances of the
pattern.

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>