Refine constant folding size limit heuristic (#2025)
Refine the size-limit heuristics used to control constant-folding. This
refinement allows some common cases to be handled automatically, such as
Transpose(weight) which is typically generated by the exporter. The
refinement looks at the increase in model-size that would be caused
replacing a node by a constant, by accounting for inputs of the node
that would be eliminated as a result of the replacement.