Squeeze Reshape Identity optimization (#2083)
A recent fix to the translation of pytorch symints introduces a
Squeeze=>Reshape pattern that can be optimized away. This PR introduces
a rewrite-rule to do this optimization.

TODO (in a separate PR): for now, this optimization needs to be
explicitly invoked. This should be done by default. (But there are
several other such optimizations that need to be collected and included
in the default-rule list.)