A couple of extensions to MHA fusion (#2106)
A couple of extensions to MHA fusion:

* One deals with variations in positions-ids. The challenge is to verify
that the position-ids used in the two RotaryEmbedding are the same. In
some models, they are the same value (by reference). In some models,
there is some duplication of the code in computing the 2D position-id
from 1D position-id. If we had a common-sub-expression
identification/elimination, that would help. For now, just handling it
in the pattern itself.

* The second deals with variations in how the last two axes of key are
transposed. Some models reshape the input tensor to 3D and do the
transpose, while some directly transpose a 4D tensor.

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>