Turn inliner into a pass and use it in rewriter & optimizer (#2149)
Use passes in optimizer and rewriter.

1. By opting into using the pass infra early, we get the benefit of
getting the additional features in pass infra w/o having to pay higher
refactoring cost in the future. We will be able to add more
sophisticated debug utilities/snapshot capabilities etc. to the passes.
2. Since we are offering the pass infra to users, we can start
validating it internally by using it here. If order altering becomes a
valid use case we can expect users may need that and we can create
appropriate facilities to support the usage.