[IR] Expose the Tape module (#2127)
Expose the `Tape` class under `ir.tape` for simplifying graph
construction in the IR. This is a secondary API for convenience. I
updated `onnxscript/ir/passes/common/shape_inference_test.py` to
demonstrate usage.

I added an optional reference to the graph from `Tape`. When the graph
is specified, the added nodes are appended to the graph. This provides
users the ability to examine the graph as they build it up using Tape.