[pass] Create check to ensure in-place passes are actually inplace (#2192)
This pull request introduces validation to ensure the in-place property
of a pass is respected.

Validation of in-place property:

*
[`onnxscript/ir/passes/_pass_infra.py`](diffhunk://#diff-70c7e5b3422f4daaf1611d4f76578c96e4c5894cced3d51718efa0290219f7f5R139-R152):
Added checks to ensure that if a pass is declared in-place, the returned
model must be the same object as the input model, and if not in-place,
the returned model must be a different object. Raises `PassError` if
these conditions are not met.