[IR] Implement `model.graphs()` (#2200)
Implement model.graphs() as a way to retrieve the main graph and all
subgraphs of it in the model.

Given

(1) how useful the method is
(2) I couldn't find an appropriate name for it in `traversal.py`
(3) Users familiar with onnxruntime optimization tools expect this
method. In PyTorch a similar `modules()` method exists.

I created this method as a core method instead of an iterator in
`traversal.py`.

Depends on https://github.com/microsoft/onnxscript/pull/2183