[IR][fix] Save value info for initializers (#1552)
Previously initializers are not included in the graph value_info because
they are not easily accessible from the Graph object. Now what we store
all the Values for initializers, we can serialize the value information
into the graph.

Updated test models to include the value info protos for initializers so
the round tripping tests can pass.

Fix https://github.com/microsoft/onnxscript/issues/1501