Add support for a non-backtracking version of pattern disjunction (#2242)
Several fusions need to support multiple variants of a pattern (such as
the optional presence of an Add or some such op). This PR adds support
for a non-backtracking version of pattern disjunction. We can now use an
"Or" between variants such as "Add(...)" and "MatMul(...)", for example.

Supporting unrestricted Or patterns is more complicated, since failure
of one alternative will require backtracking, which will require
unbinding any bindings added during the unsuccessful partial search. (We
can consider that later, if it seems useful.)