Fix attribute handling in autocast (#2256)
In https://github.com/microsoft/onnxscript/pull/2091, the call to helper
was replaced with onnx ir attribute convertion. This was not properly
handling when the attribute is a subgraph and when it uses values from
the parent scope, which the ir doesn't have access to. The IR thus
raises warnings of it not being able to find those values.

This PR reverts the change.