Returning choice values in patterns (#2284)
A fix for the use of Choice values (i.e., `OrValue([choice1, choice2])`
as a return-value in the pattern (that is, as a root of the pattern).
This is a bit complicated to support with the current implementation,
which is oriented towards iterating over the nodes in the graph, and
matching them against pattern-nodes. This PR provides a limited
extension (which handles the case in PR 2277): when the returned
choice-values are already covered by the other output values, they can
be supported by the existing matcher.

Also, on a related note, fix how value-bindings are handled in the
pattern-matcher to make it easier to return these values as the outputs
of the pattern.