Remove the RefAttr class (#2328)
## Rational

We defined the class `RefAttr` in the IR to represent reference
attributes in ONNX. Node attributes can be `Attr` and `RefAttr`.
However, since most of the time we are working with concrete attributes,
the union of types creates a typing situation where we always need to
assert the types before taking the values, even if we know a `RefAttr`
cannot exist (outside of a function definition).

This additionally matches the definition of AttributeProto in ONNX.

## Change

This change merged the two classes, and instead defines a `is_ref()`
method for users to check the reference attribute.

The change is BC breaking for usage like `isinstance(attr, ir.RefAttr)`.
Fortunately all such usages exist in this code base and not in PyTorch,
so we are safe to complete the change.