Optimize causal mask shape (#2325)
The generation of the causal mask's shape (produced by the translation
of scalar_dot_product_attention) interferes with the subsequent fusion
optimizations (because it makes use of the shape of the intermediate
matmul value).

This PR introduces a very specific fusion/rewrite to eliminate this
redundant computation of the "sequence length" dimension.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>