Fix cross attention in MHA (#2337)
Fix a seeming bug in handling of cross-attention in MHA (to be
verified): In MHA fusion, we start with an input graph where attention
is applied to 4D query/key/value, and it is transformed into a MHA op on
3D query/key/value.

In the case of cross-attention (with no rotary-embedding): the fusion
seems to convert just query to 3D, and seems to leave key and value as
4D, which seems wrong.

This PR adds the necessary 4D=>3D conversion for key/value before MHA.

Note: This is a quick fix for the relevant case (that shows up). Other
combinations may be worth checking out separately.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>